Refactor workspace setup into dedicated components
