+++ client/app/shell/SidebarLayout.tsx
@@ -7,7 +7,7 @@ import { Link, useLocation } from 'wouter'
 import { useReorderWorkspaces, useWorkspaces } from '@/client/features/home/api'
 import { workspaceKeys } from '@/client/api/workspace-keys'
 import { useWorkspaceEvent } from '@/client/runtime/useWorkspaceEvents'
-import { CreateWorkspaceDialog } from '@/client/features/home/CreateWorkspaceDialog'
+import { CreateWorkspaceDialog } from '@/client/features/home/workspace-setup/CreateWorkspaceDialog'
 import { ReorderableList } from '@/client/components/shared/ReorderableList'
 import type { ReorderableRenderState } from '@/client/components/shared/ReorderableList'
 import { Button, buttonVariants } from '@/client/components/ui/button'
